Experiment with variable pitch fonts.
In fact, writing this commit message is practically the first time I try it, and
it is ... strange. But kind of nice too. Yeah. It feels more ... civilised in a
way, more refined. And it /is/ easier to read. Except sometimes you expect too
much, such as stuff in your commit buffer turning italic when you surround it
with slashes. Oh well.