Move org faces into use-package config section
I'm not sure this is the best way, but the thinking is that I'd like the faces
to be set only after org-mode has been loaded, or else it might fail. And I'm
moving the fast tag selections away from the use-package init into a separate
section as it's pretty big and I'd like to comment on it separately.